make TestClient public to allow mocking and overriding baseURL
